requirements.txt
pymongo
PyYAML